STT_service/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
assemblyai==0.43.1
certifi==2025.8.3
click==8.2.1
fastapi==0.116.1
grpcio==1.75.0
grpcio-tools==1.75.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
websockets==15.0.1
